update the travis/karma config
