--- test/karma/chrome.js
@@ -1,8 +1,0 @@
-var common = require('./common');
-
-module.exports = function(config) {
-  config.set(common({
-    plugins: ['karma-chrome-launcher'],
-    browsers: ['Chrome']
-  }));
-};
